seperate /status & add login login to frontend
